+++ ignite/cmd/network_chain_publish.go
@@ -8,8 +8,11 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/tendermint/spn/pkg/chainid"
+	campaigntypes "github.com/tendermint/spn/x/campaign/types"
 
-	"github.com/ignite-hq/cli/ignite/pkg/clispinner"
+	"github.com/ignite-hq/cli/ignite/pkg/cliui"
+	"github.com/ignite-hq/cli/ignite/pkg/cliui/icons"
+	"github.com/ignite-hq/cli/ignite/pkg/cosmosutil"
 	"github.com/ignite-hq/cli/ignite/pkg/xurl"
 	"github.com/ignite-hq/cli/ignite/services/network"
 	"github.com/ignite-hq/cli/ignite/services/network/networkchain"
@@ -21,6 +24,7 @@ const (
 	flagHash         = "hash"
 	flagGenesis      = "genesis"
 	flagCampaign     = "campaign"
+	flagShares       = "shares"
 	flagNoCheck      = "no-check"
 	flagChainID      = "chain-id"
 	flagMainnet      = "mainnet"
@@ -46,6 +50,7 @@ func NewNetworkChainPublish() *cobra.Command {
 	c.Flags().Bool(flagNoCheck, false, "Skip verifying chain's integrity")
 	c.Flags().String(flagCampaignMetadata, "", "Add a campaign metadata")
 	c.Flags().String(flagCampaignTotalSupply, "", "Add a total of the mainnet of a campaign")
+	c.Flags().String(flagShares, "", "Add shares for the campaign")
 	c.Flags().Bool(flagMainnet, false, "Initialize a mainnet campaign")
 	c.Flags().String(flagRewardCoins, "", "Reward coins")
 	c.Flags().Int64(flagRewardHeight, 0, "Last reward height")
@@ -58,8 +63,10 @@ func NewNetworkChainPublish() *cobra.Command {
 }
 
 func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
+	session := cliui.New()
+	defer session.Cleanup()
+
 	var (
-		source                    = xurl.HTTPS(args[0])
 		tag, _                    = cmd.Flags().GetString(flagTag)
 		branch, _                 = cmd.Flags().GetString(flagBranch)
 		hash, _                   = cmd.Flags().GetString(flagHash)
@@ -69,10 +76,16 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 		noCheck, _                = cmd.Flags().GetBool(flagNoCheck)
 		campaignMetadata, _       = cmd.Flags().GetString(flagCampaignMetadata)
 		campaignTotalSupplyStr, _ = cmd.Flags().GetString(flagCampaignTotalSupply)
+		sharesStr, _              = cmd.Flags().GetString(flagShares)
 		isMainnet, _              = cmd.Flags().GetBool(flagMainnet)
 		rewardCoinsStr, _         = cmd.Flags().GetString(flagRewardCoins)
 		rewardDuration, _         = cmd.Flags().GetInt64(flagRewardHeight)
 	)
+
+	source, err := xurl.MightHTTPS(args[0])
+	if err != nil {
+		return fmt.Errorf("invalid source url format: %w", err)
+	}
 
 	if campaign != 0 && campaignTotalSupplyStr != "" {
 		return fmt.Errorf("%s and %s flags cannot be set together", flagCampaign, flagCampaignTotalSupply)
@@ -116,11 +129,10 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("%s and %s flags must be provided together", flagRewardCoins, flagRewardHeight)
 	}
 
-	nb, err := newNetworkBuilder(cmd)
+	nb, err := newNetworkBuilder(cmd, CollectEvents(session.EventBus()))
 	if err != nil {
 		return err
 	}
-	defer nb.Cleanup()
 
 	// use source from chosen target.
 	var sourceOption networkchain.SourceOption
@@ -152,12 +164,7 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 
 	initOptions = append(initOptions, networkchain.WithHome(homeDir))
 
-	// init the chain.
-	c, err := nb.Chain(sourceOption, initOptions...)
-	if err != nil {
-		return err
-	}
-
+	// prepare publish options
 	publishOptions := []network.PublishOption{network.WithMetadata(campaignMetadata)}
 
 	if genesisURL != "" {
@@ -189,14 +196,28 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 		publishOptions = append(publishOptions, network.WithTotalSupply(totalSupply))
 	}
 
+	if sharesStr != "" {
+		coins, err := cosmosutil.ParseCoinsNormalizedWithPercentageRequired(sharesStr)
+		if err != nil {
+			return err
+		}
+
+		publishOptions = append(publishOptions, network.WithShares(campaigntypes.NewSharesFromCoins(coins)))
+	}
+
+	// init the chain.
+	c, err := nb.Chain(sourceOption, initOptions...)
+	if err != nil {
+		return err
+	}
+
 	if noCheck {
 		publishOptions = append(publishOptions, network.WithNoCheck())
 	} else if err := c.Init(cmd.Context()); err != nil { // initialize the chain for checking.
 		return err
 	}
 
-	nb.Spinner.SetText("Publishing...")
-	nb.Spinner.Start()
+	session.StartSpinner("Publishing...")
 
 	n, err := nb.Network()
 	if err != nil {
@@ -214,13 +235,12 @@ func networkChainPublishHandler(cmd *cobra.Command, args []string) error {
 		}
 	}
 
-	nb.Spinner.Stop()
-
-	fmt.Printf("%s Network published \n", clispinner.OK)
-	fmt.Printf("%s Launch ID: %d \n", clispinner.Bullet, launchID)
-	fmt.Printf("%s Campaign ID: %d \n", clispinner.Bullet, campaignID)
+	session.StopSpinner()
+	session.Printf("%s Network published \n", icons.OK)
+	session.Printf("%s Launch ID: %d \n", icons.Bullet, launchID)
+	session.Printf("%s Campaign ID: %d \n", icons.Bullet, campaignID)
 	if isMainnet {
-		fmt.Printf("%s Mainnet ID: %d \n", clispinner.Bullet, mainnetID)
+		session.Printf("%s Mainnet ID: %d \n", icons.Bullet, mainnetID)
 	}
 
 	return nil