+++ ignite/pkg/cosmosgen/cosmosgen.go
@@ -2,10 +2,12 @@ package cosmosgen
 
 import (
 	"context"
+	"path/filepath"
 
 	gomodmodule "golang.org/x/mod/module"
 
 	"github.com/ignite-hq/cli/ignite/pkg/cosmosanalysis/module"
+	"github.com/ignite-hq/cli/ignite/pkg/gomodulepath"
 )
 
 // generateOptions used to configure code generation.
@@ -26,6 +28,9 @@ type generateOptions struct {
 
 // TODO add WithInstall.
 
+// ModulePathFunc defines a function type that returns a path based on a Cosmos SDK module.
+type ModulePathFunc func(module.Module) string
+
 // Option configures code generation.
 type Option func(*generateOptions)
 
@@ -33,7 +38,7 @@ type Option func(*generateOptions)
 // retrieve the path that should be used to place generated js code inside for a given module.
 // if includeThirdPartyModules set to true, code generation will be made for the 3rd party modules
 // used by the app -including the SDK- as well.
-func WithJSGeneration(includeThirdPartyModules bool, out func(module.Module) (path string)) Option {
+func WithJSGeneration(includeThirdPartyModules bool, out ModulePathFunc) Option {
 	return func(o *generateOptions) {
 		o.jsOut = out
 		o.jsIncludeThirdParty = includeThirdPartyModules
@@ -43,7 +48,7 @@ func WithJSGeneration(includeThirdPartyModules bool, out func(module.Module) (pa
 // WithVuexGeneration adds Vuex code generation. storeRootPath is used to determine the root path of generated
 // Vuex stores. includeThirdPartyModules and out configures the underlying JS lib generation which is
 // documented in WithJSGeneration.
-func WithVuexGeneration(includeThirdPartyModules bool, out func(module.Module) (path string), storeRootPath string) Option {
+func WithVuexGeneration(includeThirdPartyModules bool, out ModulePathFunc, storeRootPath string) Option {
 	return func(o *generateOptions) {
 		o.jsOut = out
 		o.jsIncludeThirdParty = includeThirdPartyModules
@@ -51,7 +56,7 @@ func WithVuexGeneration(includeThirdPartyModules bool, out func(module.Module) (
 	}
 }
 
-func WithDartGeneration(includeThirdPartyModules bool, out func(module.Module) (path string), rootPath string) Option {
+func WithDartGeneration(includeThirdPartyModules bool, out ModulePathFunc, rootPath string) Option {
 	return func(o *generateOptions) {
 		o.dartOut = out
 		o.dartIncludeThirdParty = includeThirdPartyModules
@@ -141,4 +146,13 @@ func Generate(ctx context.Context, appPath, protoDir string, options ...Option) 
 
 	return nil
 
+}
+
+// VuexStoreModulePath generates Vuex store module paths for Cosmos SDK modules.
+// The root path is used as prefix for the generated paths.
+func VuexStoreModulePath(rootPath string) ModulePathFunc {
+	return func(m module.Module) string {
+		appModulePath := gomodulepath.ExtractAppPath(m.GoModulePath)
+		return filepath.Join(rootPath, appModulePath, m.Pkg.Name, "module")
+	}
 }