+++ ignite/pkg/xurl/xurl.go
@@ -1,41 +1,76 @@
 package xurl
 
 import (
+	"errors"
 	"fmt"
+	"net"
 	"net/url"
 	"strings"
 )
 
-// TCP unsures that s url contains TCP protocol identifier.
-func TCP(s string) string {
-	if strings.HasPrefix(s, "tcp") {
-		return s
+const (
+	schemeTCP   = "tcp"
+	schemeHTTP  = "http"
+	schemeHTTPS = "https"
+	schemeWS    = "ws"
+)
+
+// TCP ensures that a URL contains a TCP scheme.
+func TCP(s string) (string, error) {
+	u, err := parseURL(s)
+	if err != nil {
+		return "", err
 	}
-	return "tcp://" + Address(s)
+
+	u.Scheme = schemeTCP
+
+	return u.String(), nil
 }
 
-// HTTP unsures that s url contains HTTP protocol identifier.
-func HTTP(s string) string {
-	if strings.HasPrefix(s, "http") {
-		return s
+// HTTP ensures that a URL contains an HTTP scheme.
+func HTTP(s string) (string, error) {
+	u, err := parseURL(s)
+	if err != nil {
+		return "", err
 	}
-	return "http://" + Address(s)
+
+	u.Scheme = schemeHTTP
+
+	return u.String(), nil
 }
 
-// HTTPS unsures that s url contains HTTPS protocol identifier.
-func HTTPS(s string) string {
-	if strings.HasPrefix(s, "https") {
-		return s
+// HTTPS ensures that a URL contains an HTTPS scheme.
+func HTTPS(s string) (string, error) {
+	u, err := parseURL(s)
+	if err != nil {
+		return "", err
 	}
-	return "https://" + Address(s)
+
+	u.Scheme = schemeHTTPS
+
+	return u.String(), nil
 }
 
-// WS unsures that s url contains WS protocol identifier.
-func WS(s string) string {
-	if strings.HasPrefix(s, "ws") {
-		return s
+// MightHTTPS ensures that a URL contains an HTTPS scheme when the current scheme is not HTTP.
+// When the URL contains an HTTP scheme it is not modified.
+func MightHTTPS(s string) (string, error) {
+	if strings.HasPrefix(strings.ToLower(s), "http://") {
+		return s, nil
 	}
-	return "ws://" + Address(s)
+
+	return HTTPS(s)
+}
+
+// WS ensures that a URL contains a WS scheme.
+func WS(s string) (string, error) {
+	u, err := parseURL(s)
+	if err != nil {
+		return "", err
+	}
+
+	u.Scheme = schemeWS
+
+	return u.String(), nil
 }
 
 // HTTPEnsurePort ensures that url has a port number suits with the connection type.
@@ -47,7 +82,7 @@ func HTTPEnsurePort(s string) string {
 
 	port := "80"
 
-	if u.Scheme == "https" {
+	if u.Scheme == schemeHTTPS {
 		port = "443"
 	}
 
@@ -56,19 +91,7 @@ func HTTPEnsurePort(s string) string {
 	return u.String()
 }
 
-// CleanPath cleans path from the url.
-func CleanPath(s string) string {
-	u, err := url.Parse(s)
-	if err != nil {
-		return s
-	}
-
-	u.Path = ""
-
-	return u.String()
-}
-
-// Address unsures that address contains localhost as host if non specified.
+// Address ensures that address contains localhost as host if non specified.
 func Address(address string) string {
 	if strings.HasPrefix(address, ":") {
 		return "localhost" + address
@@ -76,20 +99,40 @@ func Address(address string) string {
 	return address
 }
 
-// IsLocalPath checks if given address is a local fs path or a URL.
-func IsLocalPath(address string) bool {
-	for _, pattern := range []string{
-		"http://",
-		"https://",
-		"git@",
-	} {
-		if strings.HasPrefix(address, pattern) {
-			return false
-		}
-	}
-	return true
-}
-
 func IsHTTP(address string) bool {
 	return strings.HasPrefix(address, "http")
+}
+
+func parseURL(s string) (*url.URL, error) {
+	if s == "" {
+		return nil, errors.New("url is empty")
+	}
+
+	// Handle the case where the URI is an IP:PORT or HOST:PORT
+	// without scheme prefix because that case can't be URL parsed.
+	// When the URI has not scheme it is parsed as a path by "url.Parse"
+	// placing the colon within the path, which is invalid.
+	if host, isAddrPort := addressPort(s); isAddrPort {
+		return &url.URL{Host: host}, nil
+	}
+
+	p, err := url.Parse(Address(s))
+	return p, err
+}
+
+func addressPort(s string) (string, bool) {
+	// Check that the value doesn't contain a URI path
+	if strings.Index(s, "/") != -1 {
+		return "", false
+	}
+
+	// Use the net split function to support IPv6 addresses
+	host, port, err := net.SplitHostPort(s)
+	if err != nil {
+		return "", false
+	}
+	if host == "" {
+		host = "0.0.0.0"
+	}
+	return net.JoinHostPort(host, port), true
 }