+++ ignite/templates/typed/list/stargate.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/gobuffalo/genny"
 
+	"github.com/ignite-hq/cli/ignite/pkg/gomodulepath"
 	"github.com/ignite-hq/cli/ignite/pkg/placeholder"
 	"github.com/ignite-hq/cli/ignite/pkg/xgenny"
 	"github.com/ignite-hq/cli/ignite/templates/typed"
@@ -230,19 +231,19 @@ func protoQueryModify(replacer placeholder.Replacer, opts *typed.Options) genny.
 		// RPC service
 		templateRPC := `// Queries a %[2]v by id.
 	rpc %[2]v(QueryGet%[2]vRequest) returns (QueryGet%[2]vResponse) {
-		option (google.api.http).get = "/%[3]v/%[4]v/%[5]v/%[6]v/{id}";
+		option (google.api.http).get = "/%[3]v/%[4]v/%[5]v/{id}";
 	}
 
 	// Queries a list of %[2]v items.
 	rpc %[2]vAll(QueryAll%[2]vRequest) returns (QueryAll%[2]vResponse) {
-		option (google.api.http).get = "/%[3]v/%[4]v/%[5]v/%[6]v";
+		option (google.api.http).get = "/%[3]v/%[4]v/%[5]v";
 	}
 
 %[1]v`
+		appModulePath := gomodulepath.ExtractAppPath(opts.ModulePath)
 		replacementRPC := fmt.Sprintf(templateRPC, typed.Placeholder2,
 			opts.TypeName.UpperCamel,
-			opts.OwnerName,
-			opts.AppName,
+			appModulePath,
 			opts.ModuleName,
 			opts.TypeName.Snake,
 		)
@@ -395,11 +396,11 @@ func frontendSrcStoreAppModify(replacer placeholder.Replacer, opts *typed.Option
 		if err != nil {
 			return err
 		}
+		appModulePath := gomodulepath.ExtractAppPath(opts.ModulePath)
 		replacement := fmt.Sprintf(`%[1]v
-		<SpType modulePath="%[2]v.%[3]v.%[4]v" moduleType="%[5]v"  />`,
+		<SpType modulePath="%[2]v.%[3]v" moduleType="%[4]v"  />`,
 			typed.Placeholder4,
-			opts.OwnerName,
-			opts.AppName,
+			strings.ReplaceAll(appModulePath, "/", "."),
 			opts.ModuleName,
 			opts.TypeName.UpperCamel,
 		)