+++ integration/app/cmd_ibc_test.go
@@ -15,12 +15,12 @@ func TestCreateModuleWithIBC(t *testing.T) {
 
 	var (
 		env  = envtest.New(t)
-		path = env.Scaffold("blogibc")
+		path = env.Scaffold("github.com/test/blogibc")
 	)
 
 	env.Must(env.Exec("create an IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "module", "foo", "--ibc", "--require-registration"),
+			step.Exec(envtest.IgniteApp, "s", "module", "--yes", "foo", "--ibc", "--require-registration"),
 			step.Workdir(path),
 		)),
 	))
@@ -30,6 +30,7 @@ func TestCreateModuleWithIBC(t *testing.T) {
 			step.Exec(envtest.IgniteApp,
 				"s",
 				"module",
+				"--yes",
 				"appPath",
 				"--ibc",
 				"--require-registration",
@@ -42,7 +43,7 @@ func TestCreateModuleWithIBC(t *testing.T) {
 
 	env.Must(env.Exec("create a type in an IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "list", "user", "email", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "list", "--yes", "user", "email", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
@@ -53,6 +54,7 @@ func TestCreateModuleWithIBC(t *testing.T) {
 				envtest.IgniteApp,
 				"s",
 				"module",
+				"--yes",
 				"orderedfoo",
 				"--ibc",
 				"--ordering",
@@ -69,6 +71,7 @@ func TestCreateModuleWithIBC(t *testing.T) {
 				envtest.IgniteApp,
 				"s",
 				"module",
+				"--yes",
 				"unorderedfoo",
 				"--ibc",
 				"--ordering",
@@ -81,7 +84,7 @@ func TestCreateModuleWithIBC(t *testing.T) {
 
 	env.Must(env.Exec("create a non IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "module", "non_ibc", "--require-registration"),
+			step.Exec(envtest.IgniteApp, "s", "module", "--yes", "non_ibc", "--require-registration"),
 			step.Workdir(path),
 		)),
 	))
@@ -92,6 +95,7 @@ func TestCreateModuleWithIBC(t *testing.T) {
 				envtest.IgniteApp,
 				"s",
 				"module",
+				"--yes",
 				"example_with_dep",
 				"--ibc",
 				"--dep",
@@ -109,12 +113,12 @@ func TestCreateIBCOracle(t *testing.T) {
 
 	var (
 		env  = envtest.New(t)
-		path = env.Scaffold("ibcoracle")
+		path = env.Scaffold("github.com/test/ibcoracle")
 	)
 
 	env.Must(env.Exec("create an IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "module", "foo", "--ibc", "--require-registration"),
+			step.Exec(envtest.IgniteApp, "s", "module", "--yes", "foo", "--ibc", "--require-registration"),
 			step.Workdir(path),
 		)),
 	))
@@ -125,6 +129,7 @@ func TestCreateIBCOracle(t *testing.T) {
 				envtest.IgniteApp,
 				"s",
 				"module",
+				"--yes",
 				"paramsFoo",
 				"--ibc",
 				"--params",
@@ -137,21 +142,21 @@ func TestCreateIBCOracle(t *testing.T) {
 
 	env.Must(env.Exec("create the first BandChain oracle integration",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "band", "oracleone", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "band", "--yes", "oracleone", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("create the second BandChain oracle integration",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "band", "oracletwo", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "band", "--yes", "oracletwo", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("should prevent creating a BandChain oracle with no module specified",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "band", "invalidOracle"),
+			step.Exec(envtest.IgniteApp, "s", "band", "--yes", "invalidOracle"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),
@@ -159,7 +164,7 @@ func TestCreateIBCOracle(t *testing.T) {
 
 	env.Must(env.Exec("should prevent creating a BandChain oracle in a non existent module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "band", "invalidOracle", "--module", "nomodule"),
+			step.Exec(envtest.IgniteApp, "s", "band", "--yes", "invalidOracle", "--module", "nomodule"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),
@@ -167,14 +172,14 @@ func TestCreateIBCOracle(t *testing.T) {
 
 	env.Must(env.Exec("create a non-IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "module", "bar", "--params", "name,minLaunch:uint,maxLaunch:int", "--require-registration"),
+			step.Exec(envtest.IgniteApp, "s", "module", "--yes", "bar", "--params", "name,minLaunch:uint,maxLaunch:int", "--require-registration"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("should prevent creating a BandChain oracle in a non IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "band", "invalidOracle", "--module", "bar"),
+			step.Exec(envtest.IgniteApp, "s", "band", "--yes", "invalidOracle", "--module", "bar"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),
@@ -187,12 +192,12 @@ func TestCreateIBCPacket(t *testing.T) {
 
 	var (
 		env  = envtest.New(t)
-		path = env.Scaffold("blogibc2")
+		path = env.Scaffold("github.com/test/blogibc2")
 	)
 
 	env.Must(env.Exec("create an IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "module", "foo", "--ibc", "--require-registration"),
+			step.Exec(envtest.IgniteApp, "s", "module", "--yes", "foo", "--ibc", "--require-registration"),
 			step.Workdir(path),
 		)),
 	))
@@ -203,6 +208,7 @@ func TestCreateIBCPacket(t *testing.T) {
 				envtest.IgniteApp,
 				"s",
 				"packet",
+				"--yes",
 				"bar",
 				"text",
 				"texts:strings",
@@ -217,7 +223,7 @@ func TestCreateIBCPacket(t *testing.T) {
 
 	env.Must(env.Exec("should prevent creating a packet with no module specified",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "bar", "text"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "bar", "text"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),
@@ -225,7 +231,7 @@ func TestCreateIBCPacket(t *testing.T) {
 
 	env.Must(env.Exec("should prevent creating a packet in a non existent module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "bar", "text", "--module", "nomodule"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "bar", "text", "--module", "nomodule"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),
@@ -233,7 +239,7 @@ func TestCreateIBCPacket(t *testing.T) {
 
 	env.Must(env.Exec("should prevent creating an existing packet",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "bar", "post", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "bar", "post", "--module", "foo"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),
@@ -244,6 +250,7 @@ func TestCreateIBCPacket(t *testing.T) {
 			step.Exec(envtest.IgniteApp,
 				"s",
 				"packet",
+				"--yes",
 				"ticket",
 				"numInt:int",
 				"numsInt:array.int",
@@ -266,42 +273,42 @@ func TestCreateIBCPacket(t *testing.T) {
 
 	env.Must(env.Exec("create a custom field type",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "type", "custom-type", "customField:uint", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "type", "--yes", "custom-type", "customField:uint", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("create a packet with a custom field type",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "foo-baz", "customField:CustomType", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "foo-baz", "customField:CustomType", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("create a packet with no send message",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "nomessage", "foo", "--no-message", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "nomessage", "foo", "--no-message", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("create a packet with no field",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "empty", "--module", "foo"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "empty", "--module", "foo"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("create a non-IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "module", "bar", "--require-registration"),
+			step.Exec(envtest.IgniteApp, "s", "module", "--yes", "bar", "--require-registration"),
 			step.Workdir(path),
 		)),
 	))
 
 	env.Must(env.Exec("should prevent creating a packet in a non IBC module",
 		step.NewSteps(step.New(
-			step.Exec(envtest.IgniteApp, "s", "packet", "foo", "text", "--module", "bar"),
+			step.Exec(envtest.IgniteApp, "s", "packet", "--yes", "foo", "text", "--module", "bar"),
 			step.Workdir(path),
 		)),
 		envtest.ExecShouldError(),